@@ -1,10 +1,12 @@
 ﻿using System;
 using System.Diagnostics.CodeAnalysis;
 using System.Linq;
-using System.Net.Http.Headers;
 using System.Threading.Tasks;
 using Microsoft.Graph;
+using Microsoft.Graph.Models;
+using Microsoft.Graph.Users.Item.SendMail;
 using Microsoft.Identity.Client;
+using Microsoft.Kiota.Abstractions.Authentication;
 
 using t = System.Threading.Tasks;
 
@@ -37,22 +39,20 @@ namespace Facteur.MsGraph
         /// <returns></returns>
         public async t.Task SendMailAsync(EmailRequest request)
         {
-            GraphServiceClient graphClient = await ConnectClient().ConfigureAwait(false);
+            GraphServiceClient graphClient = await ConnectClient();
             Message message = new()
             {
                 Subject = request.Subject,
                 Body = new ItemBody { ContentType = BodyType.Html, Content = request.Body },
-                ToRecipients = request.To.Select(x => new Recipient { EmailAddress = new EmailAddress { Address = x } }),
-                CcRecipients = request.Cc.Select(x => new Recipient { EmailAddress = new EmailAddress { Address = x } }),
-                BccRecipients = request.Bcc.Select(x => new Recipient { EmailAddress = new EmailAddress { Address = x } }),
-                Attachments = request.AddAttachments()
+                ToRecipients = request.To.Select(x => new Recipient { EmailAddress = new EmailAddress { Address = x } }).ToList(),
+                CcRecipients = request.Cc.Select(x => new Recipient { EmailAddress = new EmailAddress { Address = x } }).ToList(),
+                BccRecipients = request.Bcc.Select(x => new Recipient { EmailAddress = new EmailAddress { Address = x } }).ToList(),
+                Attachments = request.AddAttachments().ToList()
             };
 
-            await graphClient.Users[Credentials.From]
-                .SendMail(message, false)
-                .Request()
-                .PostAsync()
-                .ConfigureAwait(false);
+            SendMailPostRequestBody requestBody = new() { Message = message };
+
+            await graphClient.Users[Credentials.From].SendMail.PostAsync(requestBody);
         }
 
         public async Task SendMailAsync(Func<IEmailComposer, Task<EmailRequest>> compose)
@@ -70,13 +70,10 @@ namespace Facteur.MsGraph
                 .WithAuthority(authority)
                 .Build();
 
-            AuthenticationResult authenticationResult = await confidentialClientApplication.AcquireTokenForClient(scopes).ExecuteAsync().ConfigureAwait(false);
-
-            return new GraphServiceClient(new DelegateAuthenticationProvider(x =>
-            {
-                x.Headers.Authorization = new AuthenticationHeaderValue("Bearer", authenticationResult.AccessToken);
-                return t.Task.FromResult(0);
-            }));
+            AuthenticationResult authenticationResult = await confidentialClientApplication.AcquireTokenForClient(scopes).ExecuteAsync();
+            TokenProvider provider = new(authenticationResult.AccessToken);
+            BaseBearerTokenAuthenticationProvider authenticationProvider = new(provider);
+            return new GraphServiceClient(authenticationProvider);
         }
     }
 }
